requirements.txt
pyyaml
numpy
scikit-learn
Pillow
pytest
